cos/queue_service.rs: colocate 15 guarantee/surplus phase tests with production (#984 P3 Phase 5c)
Append 15 tests to the existing cos/queue_service.rs::tests block:
- 5 guarantee_phase_* tests covering visit-quantum sizing + park
  policies for the guarantee phase.
- 4 surplus_phase_* tests covering weighted same-priority sharing,
  priority preference, and exact/non-exact gating.
- 4 *_guarantee_rr_* tests covering per-class round-robin cursor
  advancement.
- apply_promotion_pairs_queues_with_their_fast_path_entries +
  legacy_guarantee_rr_does_not_advance_class_cursors.

Adds apply_cos_queue_flow_fair_promotion import (cos::admission).

cargo test --bins 865/0/2.
