+++ userspace-dp/src/afxdp/cos/queue_service.rs
@@ -2214,7 +2214,567 @@ fn restore_cos_prepared_items(
 mod tests {
     use super::*;
     use crate::afxdp::tx::test_support::*;
+    use crate::afxdp::cos::admission::apply_cos_queue_flow_fair_promotion;
     use crate::afxdp::PROTO_TCP;
+
+    #[test]
+    fn surplus_phase_selects_non_exact_queue_without_guarantee_tokens() {
+        let mut root = test_cos_runtime_with_exact(false);
+        root.tokens = 1500;
+        root.queues[0].last_refill_ns = 1;
+        root.queues[0].tokens = 0;
+        root.queues[0].items.push_back(test_cos_item(1500));
+        root.queues[0].queued_bytes = 1500;
+        root.queues[0].runnable = true;
+        root.nonempty_queues = 1;
+        root.runnable_queues = 1;
+
+        assert!(select_cos_guarantee_batch(&mut root, 1).is_none());
+        let batch = select_cos_surplus_batch(&mut root, 1);
+
+        assert!(matches!(
+            batch,
+            Some(CoSBatch::Local {
+                phase: CoSServicePhase::Surplus,
+                ..
+            })
+        ));
+    }
+
+    #[test]
+    fn surplus_phase_skips_exact_queue_without_guarantee_tokens() {
+        let mut root = test_cos_runtime_with_exact(true);
+        root.tokens = 1500;
+        root.queues[0].last_refill_ns = 1;
+        root.queues[0].tokens = 0;
+        root.queues[0].items.push_back(test_cos_item(1500));
+        root.queues[0].queued_bytes = 1500;
+        root.queues[0].runnable = true;
+        root.nonempty_queues = 1;
+        root.runnable_queues = 1;
+
+        assert!(select_cos_guarantee_batch(&mut root, 1).is_none());
+        assert!(select_cos_surplus_batch(&mut root, 1).is_none());
+    }
+
+    #[test]
+    fn guarantee_phase_parks_non_exact_queue_on_root_only_wakeup() {
+        let mut root = test_cos_runtime_with_exact(false);
+        root.tokens = 0;
+        root.queues[0].last_refill_ns = 1;
+        root.queues[0].tokens = 0;
+        root.queues[0].items.push_back(test_cos_item(1500));
+        root.queues[0].queued_bytes = 1500;
+        root.queues[0].runnable = true;
+        root.nonempty_queues = 1;
+        root.runnable_queues = 1;
+
+        assert!(select_cos_guarantee_batch(&mut root, 1).is_none());
+        assert!(root.queues[0].parked);
+        assert_eq!(root.queues[0].next_wakeup_tick, 30);
+    }
+
+    #[test]
+    fn guarantee_phase_limits_service_to_visit_quantum() {
+        let mut root = test_cos_runtime_with_queues(
+            100_000_000,
+            vec![CoSQueueConfig {
+                queue_id: 0,
+                forwarding_class: "best-effort".into(),
+                priority: 5,
+                transmit_rate_bytes: 1_000_000,
+                exact: false,
+                surplus_weight: 1,
+                buffer_bytes: COS_MIN_BURST_BYTES,
+                dscp_rewrite: None,
+            }],
+        );
+        root.tokens = 64 * 1024;
+        root.queues[0].tokens = 64 * 1024;
+        root.queues[0].runnable = true;
+        for _ in 0..4 {
+            root.queues[0].items.push_back(test_cos_item(1500));
+        }
+        root.queues[0].queued_bytes = 4 * 1500;
+        root.nonempty_queues = 1;
+        root.runnable_queues = 1;
+
+        let batch = select_cos_guarantee_batch(&mut root, 1).expect("guarantee batch");
+        match batch {
+            CoSBatch::Local { items, .. } => assert_eq!(items.len(), 1),
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+        assert_eq!(root.queues[0].items.len(), 3);
+    }
+
+    #[test]
+    fn guarantee_phase_allows_larger_high_rate_visit_quantum() {
+        let mut root = test_cos_runtime_with_queues(
+            10_000_000_000u64 / 8,
+            vec![CoSQueueConfig {
+                queue_id: 0,
+                forwarding_class: "iperf-b".into(),
+                priority: 5,
+                transmit_rate_bytes: 10_000_000_000u64 / 8,
+                exact: true,
+                surplus_weight: 1,
+                buffer_bytes: 256 * 1024,
+                dscp_rewrite: None,
+            }],
+        );
+        root.tokens = 256 * 1024;
+        root.queues[0].tokens = 256 * 1024;
+        root.queues[0].runnable = true;
+        for _ in 0..200 {
+            root.queues[0].items.push_back(test_cos_item(1500));
+        }
+        root.queues[0].queued_bytes = 200 * 1500;
+        root.nonempty_queues = 1;
+        root.runnable_queues = 1;
+
+        // #920: TX_BATCH_SIZE lowered 256 → 64 caps a single visit at
+        // 64 items even when token budget would permit more (~166).
+        // The remaining tokens stay with the queue for the next visit;
+        // throughput is preserved across multiple shorter visits, with
+        // the trade-off that mouse packets get an interleave point
+        // every 64 packets instead of every 256.
+        let batch = select_cos_guarantee_batch(&mut root, 1).expect("guarantee batch");
+        match batch {
+            CoSBatch::Local { items, .. } => assert_eq!(items.len(), TX_BATCH_SIZE),
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+        assert_eq!(root.queues[0].items.len(), 200 - TX_BATCH_SIZE);
+    }
+
+    /// #920: separate from the batch-cap test above. Asserts the
+    /// rate-quantum invariant guarded by the original test name —
+    /// a 10 Gbps queue gets a strictly larger byte-budget visit
+    /// quantum than a 100 Mbps queue, regardless of TX_BATCH_SIZE.
+    /// Guards against silent regression if `cos_guarantee_quantum_bytes`
+    /// stops scaling with `transmit_rate_bytes`.
+    #[test]
+    fn guarantee_phase_quantum_scales_with_rate() {
+        // cos_guarantee_quantum_bytes is already imported at the top of tx.rs (cfg(test) block).
+        let high_rate = test_cos_runtime_with_queues(
+            10_000_000_000u64 / 8,
+            vec![CoSQueueConfig {
+                queue_id: 0,
+                forwarding_class: "iperf-b".into(),
+                priority: 5,
+                transmit_rate_bytes: 10_000_000_000u64 / 8,
+                exact: true,
+                surplus_weight: 1,
+                buffer_bytes: 256 * 1024,
+                dscp_rewrite: None,
+            }],
+        );
+        let low_rate = test_cos_runtime_with_queues(
+            100_000_000u64 / 8,
+            vec![CoSQueueConfig {
+                queue_id: 0,
+                forwarding_class: "iperf-low".into(),
+                priority: 5,
+                transmit_rate_bytes: 100_000_000u64 / 8,
+                exact: true,
+                surplus_weight: 1,
+                buffer_bytes: 256 * 1024,
+                dscp_rewrite: None,
+            }],
+        );
+        let high_q = cos_guarantee_quantum_bytes(&high_rate.queues[0]);
+        let low_q = cos_guarantee_quantum_bytes(&low_rate.queues[0]);
+        assert!(
+            high_q > low_q,
+            "high-rate quantum ({high_q}) must exceed low-rate quantum ({low_q})"
+        );
+    }
+
+    #[test]
+    fn guarantee_phase_rotates_between_backlogged_queues() {
+        let mut root = test_cos_runtime_with_queues(
+            100_000_000,
+            vec![
+                CoSQueueConfig {
+                    queue_id: 0,
+                    forwarding_class: "best-effort".into(),
+                    priority: 5,
+                    transmit_rate_bytes: 1_000_000,
+                    exact: false,
+                    surplus_weight: 1,
+                    buffer_bytes: COS_MIN_BURST_BYTES,
+                    dscp_rewrite: None,
+                },
+                CoSQueueConfig {
+                    queue_id: 1,
+                    forwarding_class: "af11".into(),
+                    priority: 5,
+                    transmit_rate_bytes: 1_000_000,
+                    exact: false,
+                    surplus_weight: 1,
+                    buffer_bytes: COS_MIN_BURST_BYTES,
+                    dscp_rewrite: None,
+                },
+            ],
+        );
+        root.tokens = 64 * 1024;
+        for queue in &mut root.queues {
+            queue.tokens = 64 * 1024;
+            queue.runnable = true;
+            queue.items.push_back(test_cos_item(1500));
+            queue.items.push_back(test_cos_item(1500));
+            queue.queued_bytes = 2 * 1500;
+        }
+        root.nonempty_queues = 2;
+        root.runnable_queues = 2;
+
+        let first = select_cos_guarantee_batch(&mut root, 1).expect("first guarantee batch");
+        let second = select_cos_guarantee_batch(&mut root, 1).expect("second guarantee batch");
+
+        match first {
+            CoSBatch::Local { queue_idx, .. } => assert_eq!(queue_idx, 0),
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+        match second {
+            CoSBatch::Local { queue_idx, .. } => assert_eq!(queue_idx, 1),
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+    }
+
+    #[test]
+    fn exact_and_nonexact_guarantee_rr_cursors_advance_independently() {
+        // #689 regression. Prior to the cursor split, serving an exact
+        // queue advanced the shared `guarantee_rr` and could cause the
+        // non-exact pass to skip a waiting queue on its next run. Pin
+        // that the exact pass does not touch `nonexact_guarantee_rr`
+        // and vice versa.
+        let mut root = test_mixed_class_root_with_primed_queues();
+        assert_eq!(root.exact_guarantee_rr, 0);
+        assert_eq!(root.nonexact_guarantee_rr, 0);
+
+        // Serving an exact queue must not disturb the non-exact cursor.
+        let selection = select_exact_cos_guarantee_queue_with_fast_path(&mut root, &[], 1)
+            .expect("exact queue selection");
+        assert_eq!(selection.queue_idx, 0);
+        assert_eq!(
+            root.exact_guarantee_rr, 1,
+            "exact cursor must advance past the served queue"
+        );
+        assert_eq!(
+            root.nonexact_guarantee_rr, 0,
+            "serving an exact queue must not advance the non-exact cursor"
+        );
+
+        // Serving a non-exact queue must not disturb the exact cursor.
+        let batch =
+            select_nonexact_cos_guarantee_batch(&mut root, 1).expect("nonexact queue batch");
+        match batch {
+            CoSBatch::Local { queue_idx, .. } => assert_eq!(queue_idx, 1),
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+        assert_eq!(
+            root.exact_guarantee_rr, 1,
+            "non-exact service must not advance the exact cursor"
+        );
+        assert_eq!(
+            root.nonexact_guarantee_rr, 2,
+            "non-exact cursor must advance past the served queue"
+        );
+    }
+
+    #[test]
+    fn exact_guarantee_rr_walks_exact_queues_in_order_independent_of_nonexact() {
+        // Exact queues must rotate exact-0 -> exact-2 -> exact-0 -> exact-2
+        // regardless of non-exact activity between calls. #689 before-fix
+        // behavior under the shared cursor was: exact-0 served (rr=1),
+        // then a non-exact service would bump rr past exact-2's position,
+        // so the next exact call would skip exact-2 and loop back to
+        // exact-0. This test pins that the split cursor rotates exact
+        // queues deterministically without regard for non-exact service.
+        // Helper primes eight 1500-byte items and sets `queued_bytes`
+        // to match; no additional priming needed here. Only bump
+        // queue.tokens on the exact queues to make sure they never hit
+        // token-starvation during the four interleaved rounds below —
+        // the exact selector does not refill exact-queue tokens itself
+        // (that is done by the shared-lease path), so this test bypasses
+        // that machinery by handing the queues a large local budget.
+        let mut root = test_mixed_class_root_with_primed_queues();
+        for queue in &mut root.queues {
+            if queue.exact {
+                queue.tokens = 128 * 1024;
+            }
+        }
+
+        let mut exact_order = Vec::new();
+        for _ in 0..4 {
+            // Interleave a non-exact service between exact calls; the exact
+            // rotation must not notice.
+            let selection = select_exact_cos_guarantee_queue_with_fast_path(&mut root, &[], 1)
+                .expect("exact queue");
+            exact_order.push(selection.queue_idx);
+            // Service a non-exact queue to simulate concurrent class activity;
+            // ignore the result.
+            let _ = select_nonexact_cos_guarantee_batch(&mut root, 1);
+        }
+        assert_eq!(exact_order, vec![0, 2, 0, 2]);
+    }
+
+    #[test]
+    fn nonexact_guarantee_rr_walks_nonexact_queues_in_order_independent_of_exact() {
+        // Symmetric to the exact test: non-exact rotation is 1 -> 3 -> 1 -> 3
+        // regardless of exact-queue activity between calls. Helper primes
+        // eight 1500-byte items per queue with `queued_bytes` already
+        // consistent; no additional priming needed.
+        let mut root = test_mixed_class_root_with_primed_queues();
+
+        let mut nonexact_order = Vec::new();
+        for _ in 0..4 {
+            let batch = select_nonexact_cos_guarantee_batch(&mut root, 1).expect("nonexact batch");
+            let queue_idx = match batch {
+                CoSBatch::Local { queue_idx, .. } => queue_idx,
+                CoSBatch::Prepared { queue_idx, .. } => queue_idx,
+            };
+            nonexact_order.push(queue_idx);
+            // Interleave an exact service; must not disturb non-exact rotation.
+            let _ = select_exact_cos_guarantee_queue_with_fast_path(&mut root, &[], 1);
+        }
+        assert_eq!(nonexact_order, vec![1, 3, 1, 3]);
+    }
+
+    #[test]
+    fn legacy_guarantee_rr_does_not_advance_class_cursors() {
+        // The entire reason `legacy_guarantee_rr` exists as a third cursor
+        // (instead of the legacy unified selector reusing one of the
+        // production cursors) is to keep the legacy walk isolated from the
+        // production exact/nonexact rotation state. Pin that contract:
+        // a call through the legacy selector must advance only its own
+        // cursor, never the two production cursors.
+        let mut root = test_mixed_class_root_with_primed_queues();
+        let batch = select_cos_guarantee_batch(&mut root, 1).expect("legacy guarantee batch");
+        // Served something, so `legacy_guarantee_rr` advanced.
+        match batch {
+            CoSBatch::Local { queue_idx, .. } => {
+                assert_eq!(queue_idx, 0, "legacy walk starts at index 0");
+            }
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+        assert_eq!(root.legacy_guarantee_rr, 1);
+        // Production cursors untouched — this is the isolation guarantee
+        // that justifies the extra field over reusing either production
+        // cursor for the legacy walk.
+        assert_eq!(
+            root.exact_guarantee_rr, 0,
+            "legacy selector must not advance exact production cursor"
+        );
+        assert_eq!(
+            root.nonexact_guarantee_rr, 0,
+            "legacy selector must not advance nonexact production cursor"
+        );
+    }
+
+    #[test]
+    fn guarantee_rr_cursors_start_at_zero_after_runtime_build() {
+        // Pin the invariant that a fresh runtime starts with both cursors
+        // at 0. `build_cos_interface_runtime` is the one production init
+        // site; any refactor that accidentally leaves a cursor uninitialized
+        // or drops one of the fields fails here.
+        let root = test_cos_runtime_with_queues(
+            10_000_000_000 / 8,
+            vec![CoSQueueConfig {
+                queue_id: 0,
+                forwarding_class: "q0".into(),
+                priority: 5,
+                transmit_rate_bytes: 1_000_000_000 / 8,
+                exact: true,
+                surplus_weight: 1,
+                buffer_bytes: COS_MIN_BURST_BYTES,
+                dscp_rewrite: None,
+            }],
+        );
+        assert_eq!(root.exact_guarantee_rr, 0);
+        assert_eq!(root.nonexact_guarantee_rr, 0);
+        assert_eq!(root.legacy_guarantee_rr, 0);
+    }
+
+    #[test]
+    fn surplus_phase_prefers_higher_priority_queue() {
+        let mut root = test_cos_runtime_with_queues(
+            100_000_000,
+            vec![
+                CoSQueueConfig {
+                    queue_id: 0,
+                    forwarding_class: "bulk".into(),
+                    priority: 5,
+                    transmit_rate_bytes: 1_000_000,
+                    exact: false,
+                    surplus_weight: 1,
+                    buffer_bytes: COS_MIN_BURST_BYTES,
+                    dscp_rewrite: None,
+                },
+                CoSQueueConfig {
+                    queue_id: 1,
+                    forwarding_class: "voice".into(),
+                    priority: 0,
+                    transmit_rate_bytes: 1_000_000,
+                    exact: false,
+                    surplus_weight: 1,
+                    buffer_bytes: COS_MIN_BURST_BYTES,
+                    dscp_rewrite: None,
+                },
+            ],
+        );
+        root.tokens = 64 * 1024;
+        for queue in &mut root.queues {
+            queue.last_refill_ns = 1;
+            queue.tokens = 0;
+            queue.runnable = true;
+            queue.items.push_back(test_cos_item(1500));
+            queue.queued_bytes = 1500;
+        }
+        root.nonempty_queues = 2;
+        root.runnable_queues = 2;
+
+        assert!(select_cos_guarantee_batch(&mut root, 1).is_none());
+        let batch = select_cos_surplus_batch(&mut root, 1).expect("surplus batch");
+        match batch {
+            CoSBatch::Local { queue_idx, .. } => assert_eq!(queue_idx, 1),
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+    }
+
+    #[test]
+    fn surplus_phase_applies_weighted_same_priority_sharing() {
+        let mut root = test_cos_runtime_with_queues(
+            100_000_000,
+            vec![
+                CoSQueueConfig {
+                    queue_id: 0,
+                    forwarding_class: "small".into(),
+                    priority: 5,
+                    transmit_rate_bytes: 1_000_000,
+                    exact: false,
+                    surplus_weight: 1,
+                    buffer_bytes: COS_MIN_BURST_BYTES,
+                    dscp_rewrite: None,
+                },
+                CoSQueueConfig {
+                    queue_id: 1,
+                    forwarding_class: "large".into(),
+                    priority: 5,
+                    transmit_rate_bytes: 4_000_000,
+                    exact: false,
+                    surplus_weight: 4,
+                    buffer_bytes: COS_MIN_BURST_BYTES,
+                    dscp_rewrite: None,
+                },
+            ],
+        );
+        root.tokens = 64 * 1024;
+        for queue in &mut root.queues {
+            queue.last_refill_ns = 1;
+            queue.tokens = 0;
+            queue.runnable = true;
+            for _ in 0..8 {
+                queue.items.push_back(test_cos_item(1500));
+            }
+            queue.queued_bytes = 8 * 1500;
+        }
+        root.nonempty_queues = 2;
+        root.runnable_queues = 2;
+
+        let first = select_cos_surplus_batch(&mut root, 1).expect("first surplus batch");
+        let second = select_cos_surplus_batch(&mut root, 1).expect("second surplus batch");
+
+        match first {
+            CoSBatch::Local {
+                queue_idx, items, ..
+            } => {
+                assert_eq!(queue_idx, 0);
+                assert_eq!(items.len(), 1);
+            }
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+        match second {
+            CoSBatch::Local {
+                queue_idx, items, ..
+            } => {
+                assert_eq!(queue_idx, 1);
+                assert_eq!(items.len(), 4);
+            }
+            CoSBatch::Prepared { .. } => panic!("expected local batch"),
+        }
+    }
+
+    /// Pin that `apply_cos_queue_flow_fair_promotion` propagates the
+    /// per-queue `shared_exact` bits correctly when the interface
+    /// has a mix of shared_exact and owner-local-exact queues — the
+    /// common production shape (a low-rate iperf-a queue next to a
+    /// high-rate iperf-c queue on the same interface). Breaking the
+    /// zip alignment between `runtime.queues` and
+    /// `iface_fast.queue_fast_path` at the
+    /// `ensure_cos_interface_runtime` call site would swap the two
+    /// queues' `shared_exact` shadows and their `flow_fair` bits,
+    /// silently routing both to the wrong admission branch and
+    /// turning off SFQ on the iperf-a queue (re-breaking #784).
+    #[test]
+    fn apply_promotion_pairs_queues_with_their_fast_path_entries() {
+        let mut runtime = test_cos_runtime_with_queues(
+            100_000_000_000 / 8,
+            vec![
+                CoSQueueConfig {
+                    queue_id: 4,
+                    forwarding_class: "iperf-a".into(),
+                    priority: 5,
+                    transmit_rate_bytes: 1_000_000_000 / 8,
+                    exact: true,
+                    surplus_weight: 1,
+                    buffer_bytes: 128 * 1024,
+                    dscp_rewrite: None,
+                },
+                CoSQueueConfig {
+                    queue_id: 5,
+                    forwarding_class: "iperf-c".into(),
+                    priority: 5,
+                    transmit_rate_bytes: 25_000_000_000 / 8,
+                    exact: true,
+                    surplus_weight: 1,
+                    buffer_bytes: 128 * 1024,
+                    dscp_rewrite: None,
+                },
+            ],
+        );
+
+        // Position 0 -> owner-local-exact; position 1 -> shared_exact.
+        let fast_path = vec![
+            test_queue_fast_path_for_promotion(false),
+            test_queue_fast_path_for_promotion(true),
+        ];
+        apply_cos_queue_flow_fair_promotion(&mut runtime, &fast_path, 0);
+
+        assert!(
+            runtime.queues[0].flow_fair,
+            "queue at position 0 (iperf-a, shared_exact=false) must \
+             be on the flow-fair path — #784 fairness fix depends on it",
+        );
+        assert!(
+            !runtime.queues[0].shared_exact,
+            "queue at position 0 must get position-0's shared_exact=false",
+        );
+        assert!(
+            runtime.queues[1].flow_fair,
+            "#785 Phase 3: queue at position 1 (iperf-c, \
+             shared_exact=true) must also be on the flow-fair path \
+             so MQFQ VFT ordering enforces per-flow fairness. The \
+             admission gates (cos_queue_flow_share_limit, \
+             apply_cos_admission_ecn_policy) separately downgrade to \
+             aggregate-only on shared_exact queues.",
+        );
+        assert!(
+            runtime.queues[1].shared_exact,
+            "queue at position 1 must get position-1's shared_exact=true \
+             — zip misalignment would silently mis-route admission policy",
+        );
+    }
+
     use crate::afxdp::types::{COS_FLOW_FAIR_BUCKETS, CoSQueueConfig, CoSQueueDropCounters, CoSQueueOwnerProfile, FlowRrRing};
 
     #[test]
